Singleton Sorts For ConstantEntry
This commit adds the implementation of SingletonSorts to the pre-model
SPARC program.  Each ConstantEntry, even if it is just a simple ID, is
given its own parent sort which is added as sub-sorts to the sorts the
constant is declared to belong to.   The singleton sorts are erased from
usage in the TM program as they are just a gimmick to get around SPARC's
sort syntax.