Simple Constant Definition
This addition adds the implementation of simple constant definitions.
Occurrences of the constants in rules are replaced with the defining
term.   Sort singletons have their simple constants replaced with their
defintions.  Semantic errors should be thrown for antyhing that is not a
simple constant definition.   

The coin_flip test program now has a substituion of heads and tails with
front and back, respectively.